requirements/development.txt
-r base.txt

# Development Tools
django-debug-toolbar==5.1.0
django-extensions==4.1
django-filter==23.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
coverage==7.3.2

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Development Server
django-cors-headers==4.3.1